fix: some ui bugs in table view
