+++ apps/core/src/bindings.ts
@@ -131,8 +131,8 @@ export const events = __makeEvents__<{
 
 /** user-defined constants **/
 
-export const SETTINGS_FILE_NAME = "settings.json" as const
 export const KEYBINDINGS_FILE_NAME = "keybindings.json" as const
+export const SETTINGS_FILE_NAME = "settings.json" as const
 
 /** user-defined types **/
