+++ apps/core/src/components/ui/sidebar.tsx
@@ -132,7 +132,7 @@ const SidebarProvider = React.forwardRef<
               } as React.CSSProperties
             }
             className={cn(
-              "group/sidebar-wrapper flex min-h-svh w-full has-[[data-variant=inset]]:bg-sidebar",
+              "group/sidebar-wrapper flex h-full w-full has-[[data-variant=inset]]:bg-sidebar",
               className
             )}
             ref={ref}
