+++ apps/core/src/lib/constants.ts
@@ -1,7 +1,8 @@
 export const QUERY_KEYS = {
   GET_TABLES: "get_tables",
   GET_TABLE_ROWS: "get_table_rows",
-  GET_CONNECTION_DETAILS: "get_connection_details"
+  GET_CONNECTION_DETAILS: "get_connection_details",
+  TABLE_COLUMNS: "table_columns"
 }
 
 export const LOCAL_STORAGE = {
