+++ apps/core/src/routes/dashboard/_layout/_table-view-layout.tsx
@@ -22,9 +22,7 @@ function TableViewLayout() {
       >
         <TableSelectionBreadCrumb connectionId={connectionId!} />
       </div>
-      <div className="flex-1">
-        <Outlet />
-      </div>
+      <Outlet />
     </div>
   )
 }
