+++ apps/core/src/routes/dashboard/_layout/_table-view-layout/table-view/$tableName/route.tsx
@@ -1,22 +1,41 @@
-import LoadingSpinner from "@/components/loading-spinner"
-import { getTablesQueryOptions } from "@/features/table-view/queries"
-import { useGetTableColumns } from "@/hooks/table"
+import { DataTable } from "@/components/custom/data-table"
+import { Skeleton } from "@/components/ui/skeleton"
+import {
+  getTableColumnsOptions,
+  getTablesQueryOptions
+} from "@/features/table-view/queries"
+import { useSetupReactTable } from "@/hooks/table"
+import { useTauriEventListener } from "@/hooks/use-tauri-event-listener"
+import { QUERY_KEYS } from "@/lib/constants"
+import { useSuspenseQuery } from "@tanstack/react-query"
 import { createFileRoute } from "@tanstack/react-router"
-import { toast } from "sonner"
-import DataTable from "./-components/data-table"
 
 export const Route = createFileRoute(
   "/dashboard/_layout/_table-view-layout/table-view/$tableName"
 )({
   loaderDeps: ({ search }) => ({ connectionId: search.connectionId }),
-  loader: ({ context: { queryClient }, deps: { connectionId } }) =>
-    queryClient.ensureQueryData(getTablesQueryOptions(connectionId!)),
-  component: TableView
+  loader: ({
+    context: { queryClient },
+    deps: { connectionId },
+    params: { tableName }
+  }) => {
+    queryClient.ensureQueryData(getTablesQueryOptions(connectionId!))
+    queryClient.ensureQueryData(getTableColumnsOptions(tableName))
+  },
+  component: TableView,
+  pendingComponent: () => (
+    <div className="space-y-2">
+      <Skeleton className="h-10 w-full" />
+      <Skeleton className="h-full" />
+    </div>
+  )
 })
 
 function TableView() {
   const { tableName } = Route.useParams()
+  const { queryClient } = Route.useRouteContext()
   const { connectionId } = Route.useSearch()
+  const { data: columns } = useSuspenseQuery(getTableColumnsOptions(tableName))
   // const { updateTableName } = useTableState()
   // const [, setActiveTable] = useLocalStorage<TableLocalStorage | null>(
   //   `@tablex/${connectionId}`,
@@ -29,24 +48,32 @@ function TableView() {
   //   [setActiveTable, tableName]
   // )
 
-  const {
-    data: columns,
-    isPending: isColumnsPending,
-    isError,
-    error
-  } = useGetTableColumns(tableName)
+  const { isRowsLoading, contextMenuRow, setContextMenuRow, table, tableRef } =
+    useSetupReactTable({ columns, tableName, connectionId })
 
-  if (isColumnsPending) return <LoadingSpinner />
+  useTauriEventListener(
+    "tableContentsChanged",
+    () =>
+      queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.GET_TABLE_ROWS] }),
+    [queryClient]
+  )
 
-  if (isError) return toast.error(error.message, { id: "get_table_columns" })
+  if (isRowsLoading)
+    return (
+      <div className="flex h-full flex-col space-y-10">
+        <Skeleton className="h-10 w-full" />
+        <Skeleton className="h-full" />
+      </div>
+    )
 
   return (
     <section className="flex h-full w-full flex-col overflow-auto will-change-scroll">
-      <DataTable
+      <DataTable table={table} />
+      {/* <DataTable
         columns={columns!}
         connectionId={connectionId}
         tableName={tableName}
-      />
+      /> */}
     </section>
   )
 }
