+++ package.json
@@ -51,6 +51,7 @@
     "postcss": "^8.5.3",
     "prettier": "3.1.1",
     "prettier-plugin-tailwindcss": "^0.6.11",
+    "react-scan": "^0.3.3",
     "release-it": "^17.11.0",
     "tailwindcss": "^3.4.17",
     "typescript": "^5.7.3"
